Add: Fuse Set: Fuse byt high , Fix: previous commite
