+++ include/mcusim/avr/sim/sim.h
@@ -108,34 +108,6 @@ enum MSIM_AVRSREGFlag {
 	AVR_SREG_T_BIT,
 	AVR_SREG_GLOB_INT
 };
-#ifdef AVR_ATMEGA328P
-/* ATMega328P Fuse Low Byte */
-enum MSIM_AVRFuseLowByte {
-	AVR_FLB_CKSEL0,
-	AVR_FLB_CKSEL1,
-	AVR_FLB_CKSEL2,
-	AVR_FLB_CKSEL3,
-	AVR_FLB_SUT0,
-	AVR_FLB_SUT1,
-	AVR_FLB_CKOUT,
-	AVR_FLB_CKDIV8
-};
-enum MSIM_AVRFuseHighByte {
-	AVR_FHB_BOOTRST,
-	AVR_FHB_BOOTSZ0,
-	AVR_FHB_BOOTSZ1,
-	AVR_FHB_EESAVE,
-	AVR_FHB_WDTON,
-	AVR_FHB_SPIEN,
-	AVR_FHB_DWEN,
-	AVR_FHB_RSTDISBL
-};
-enum MSIM_AVRFuseExtended {
-	AVR_FEXT_BODLEVEL0,
-	AVR_FEXT_BODLEVEL1,
-	AVR_FEXT_BODLEVEL2,
-};
-#endif
 /* Instance of the AVR microcontroller. */
 struct MSIM_AVR {
 	char name[20];			/* Name of the MCU */
